requirements.txt
leidenalg==0.10.2
networkx==3.4.1
numpy==2.2.1
pandas==2.2.3
python_igraph==0.11.8
PyYAML==6.0.2
scikit_learn==1.6.0
scipy==1.14.1
tqdm==4.66.5
